# doc/API.yml
openapi: 3.0.3
info:
  title: Todo Service API
  version: 1.0.0
  description: Contract-first API for managing todo items.

servers:
  - url: /api/v1
    description: Relative path (for API gateway)

paths:
  /todos:
    post:
      tags:
        - todos
      operationId: createTodo
      summary: Create a new todo item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoDto'
      responses:
        '201':
          description: Todo item created
          headers:
            Location:
              description: URL of the created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItemDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - todos
      operationId: listTodos
      summary: List todo items
      parameters:
        - in: query
          name: statuses
          required: false
          description: |
            Filter todos by status. Currently supported values: `not_done`.
            Omit or leave empty to return all items.
            Accepts multiple values for future extensibility.
          schema:
            type: array
            items:
              type: string
            default: [ ]
          style: form
          explode: true
      responses:
        '200':
          description: List of todo items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoItemDto'
  /todos/{id}:
    get:
      tags:
        - todos
      operationId: getTodoById
      summary: Get todo item by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Todo item found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItemDto'
        '404':
          description: Todo item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /todos/{id}/description:
    put:
      tags:
        - todos
      operationId: updateDescription
      summary: Update description of a todo item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDescriptionDto'
      responses:
        '200':
          description: Description updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItemDto'
        '404':
          description: Todo item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot change description of past due item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /todos/{id}/done:
    post:
      tags:
        - todos
      operationId: markDone
      summary: Mark a todo item as done
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Todo item marked as done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItemDto'
        '404':
          description: Todo item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot mark past due item as done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /todos/{id}/not-done:
    post:
      tags:
        - todos
      operationId: markNotDone
      summary: Mark a todo item as not done
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Todo item not done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItemDto'
        '404':
          description: Todo item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot undo a past due item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    TodoStatusDto:
      type: string
      description: Status of the todo item
      enum:
        - not_done
        - done
        - past_due
    TodoItemDto:
      type: object
      required:
        - id
        - description
        - status
        - createdAt
        - dueAt
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/TodoStatusDto'
        createdAt:
          type: string
          format: date-time
        dueAt:
          type: string
          format: date-time
        doneAt:
          type: string
          format: date-time
          nullable: true
    CreateTodoDto:
      type: object
      required:
        - description
        - dueAt
      properties:
        description:
          type: string
          minLength: 1
        dueAt:
          type: string
          format: date-time
    UpdateDescriptionDto:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          minLength: 1

    ErrorResponse:
      type: object
      required:
        - message
        - code
        - timestamp
      properties:
        message:
          type: string
          description: Error message describing what went wrong
        code:
          type: string
          description: Error code for programmatic handling
        timestamp:
          type: string
          format: date-time
          description: When the error occurred