# src/main/java/com/simplesystem/todoservice/api/impl/TodosController.java
package com.simplesystem.todoservice.api.impl;

import com.simplesystem.todoservice.api.TodosApi;
import com.simplesystem.todoservice.api.model.CreateTodoDto;
import com.simplesystem.todoservice.api.model.TodoItemDto;
import com.simplesystem.todoservice.api.model.UpdateDescriptionDto;
import com.simplesystem.todoservice.mapper.TodoMapper;
import com.simplesystem.todoservice.model.TodoStatus;
import com.simplesystem.todoservice.service.TodoService;
import lombok.RequiredArgsConstructor;
import lombok.val;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import java.util.List;
import java.util.stream.Collectors;

@RestController
@RequiredArgsConstructor
public class TodosController implements TodosApi {

    private final TodoService todoService;
    private final TodoMapper todoMapper;

    @Override
    public ResponseEntity<TodoItemDto> createTodo(CreateTodoDto createTodoDto) {
        val created = todoService.createTodo(createTodoDto.getDescription(), createTodoDto.getDueAt());
        val body = todoMapper.mapToDto(created);
        val location = ServletUriComponentsBuilder
                .fromCurrentRequest()
                .path("/{id}")
                .buildAndExpand(created.getId())
                .toUri();
        return ResponseEntity.created(location).body(body);
    }

    @Override
    public ResponseEntity<TodoItemDto> getTodoById(Long id) {
        val item = todoService.getTodo(id);
        return ResponseEntity.ok(todoMapper.mapToDto(item));
    }

    @Override
    public ResponseEntity<List<TodoItemDto>> listTodos(List<String> statuses) {
        val convertedStatuses = statuses.stream()
                .map(s -> TodoStatus.valueOf(s.toUpperCase()))
                .toList();
        val items = todoService.listTodos(convertedStatuses)
                .stream()
                .map(todoMapper::mapToDto)
                .collect(Collectors.toList());
        return ResponseEntity.ok(items);
    }

    @Override
    public ResponseEntity<TodoItemDto> updateDescription(Long id, UpdateDescriptionDto updateDescriptionDto) {
        val updated = todoService.updateDescription(id, updateDescriptionDto.getDescription());
        return ResponseEntity.ok(todoMapper.mapToDto(updated));
    }

    @Override
    public ResponseEntity<TodoItemDto> markDone(Long id) {
        val updated = todoService.markDone(id);
        return ResponseEntity.ok(todoMapper.mapToDto(updated));
    }

    @Override
    public ResponseEntity<TodoItemDto> markNotDone(Long id) {
        val updated = todoService.markNotDone(id);
        return ResponseEntity.ok(todoMapper.mapToDto(updated));
    }
}